src: Remove ES5/6 shim dependencies
This project is a pretty lightweight module and can be implemented
as a zero dependency library.

The actual implementation is basically 20 lines itself.

Since this module already depends on ES5 & global.Promise I
inlined the other 3 dependencies it has (all pretty trivial).

This makes this module zero dependency which makes it really
lightweight!

The motiviation for this is to simplify `aws-sdk` ;

Currently for aws-sdk; 50% of the transitive dependencies are
the dependencies of `util.promisify` itself ( 24 ). This is
quite a lot of weight.